rename package to rewriter
