+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/AllRewriteTests.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.core.rewrite;
 
 import org.apache.shardingsphere.core.rewrite.builder.SQLBuilderTest;
-import org.apache.shardingsphere.core.rewrite.engine.ShardingSQLRewriteEngineTest;
+import org.apache.shardingsphere.core.rewrite.rewriter.ShardingSQLRewriteEngineTest;
 import org.apache.shardingsphere.core.rewrite.placeholder.AllPlaceholderTests;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
